requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
